Solve lintes
