+++ data/albums.json
@@ -1,1 +1,1 @@
-[{"id":178,"on_spotify":true,"date":"02-02-2012","genre":2,"author":25,"label":1},{"id":794,"on_spotify":true,"date":"03-02-2012","genre":1,"author":25,"label":2}]
+[{"id":178,"on_spotify":true,"date":"02-02-2012","genre":2,"author":19,"label":1},{"id":794,"on_spotify":true,"date":"03-02-2012","genre":1,"author":19,"label":2},{"id":770,"on_spotify":true,"date":"02-02-2012","genre":2,"author":19,"label":3},{"id":30,"on_spotify":true,"date":"04-05-1998","genre":2,"author":19,"label":4}]
