+++ data/books.json
@@ -1,1 +1,1 @@
-[{"id":685,"publisher":"Anagrama","cover_state":"Good","genre":1,"author":25,"label":1},{"id":754,"publisher":"asddd","cover_state":"123","genre":1,"author":25,"label":1},{"id":856,"publisher":"123","cover_state":"asdg","genre":1,"author":25,"label":1},{"id":333,"publisher":"frank","cover_state":"good","genre":1,"author":25,"label":1}]
+[{"id":565,"publisher":"Anagrama","cover_state":"Good","genre":1,"author":19,"label":1},{"id":442,"publisher":"asddd","cover_state":"123","genre":1,"author":19,"label":1},{"id":264,"publisher":"123","cover_state":"asdg","genre":1,"author":19,"label":1},{"id":510,"publisher":"frank","cover_state":"good","genre":1,"author":19,"label":1},{"id":229,"publisher":"juan","cover_state":"bad","genre":1,"author":19,"label":1}]
