+++ data/genres.json
@@ -1,1 +1,1 @@
-[{"id":1,"name":"Jazz","items":[794]},{"id":2,"name":"Rock","items":[178]},{"id":3,"name":"Pop","items":[]}]
+[{"id":1,"name":"Jazz","items":[794]},{"id":2,"name":"Rock","items":[178,770,30]},{"id":3,"name":"Pop","items":[]}]
