+++ data/labels.json
@@ -1,1 +1,1 @@
-[{"id":1,"title":"Label 1","color":"red","items":[178]},{"id":2,"title":"Label 2","color":"blue","items":[794]},{"id":3,"title":"Label 3","color":"green","items":[]},{"id":4,"title":"Label 4","color":"yellow","items":[]},{"id":5,"title":"Label 5","color":"purple","items":[]}]
+[{"id":1,"title":"Label 1","color":"red","items":[178]},{"id":2,"title":"Label 2","color":"blue","items":[794]},{"id":3,"title":"Label 3","color":"green","items":[770]},{"id":4,"title":"Label 4","color":"yellow","items":[30]},{"id":5,"title":"Label 5","color":"purple","items":[]}]
