+++ modules/store.rb
@@ -1,20 +1,13 @@
 require 'json'
-require 'pry'
+require_relative 'store_basis'
 module Store
+  include StoreBasis
   def save_albums
     File.write('data/albums.json', @music_albums.map { |album| album_to_hash album }.to_json)
   end
 
   def save_books
     File.write('data/books.json', @books.map { |book| book_to_hash book }.to_json)
-  end
-
-  def save_labels
-    File.write('data/labels.json', @labels.map { |label| label_to_hash label }.to_json)
-  end
-
-  def save_genres
-    File.write('data/genres.json', @genres.map { |genre| genre_to_hash genre }.to_json)
   end
 
   def save_games
@@ -71,46 +64,7 @@ module Store
     end
   end
 
-  def load_labels
-    return unless File.exist?('data/labels.json')
-
-    puts 'Loading Labels'
-    file = File.read('data/labels.json')
-    label_data = JSON.parse(file)
-    label_data.each do |label|
-      aux_label = Label.new label['title'], label['color']
-      aux_label.add_id(label['id'])
-      if label['items'].nil? && @labels.any?
-        label['items'].each do |item|
-          aux_label.add_item(get_labels_by_id(item))
-        end
-      end
-      @labels << aux_label
-    end
-  end
-
-  def get_item_by_id(id)
-    @music_albums.find { |album| album.id == id }
-  end
-
-  def get_labels_by_id(id)
-    @books.find { |_album| book.id == id }
-  end
-
-  def load_genres
-    return unless File.exist?('data/genres.json')
-
-    puts 'loading genres'
-
-    JSON.parse(File.read('data/genres.json')).each do |genre|
-      genree = Genre.new(genre['name'])
-      genree.add_id(genre['id'])
-      @genres << genree
-    end
-  end
-
   def load_games
-    puts 'loading games'
     return unless File.exist?('./data/games.json')
 
     games = JSON.parse(File.read('./data/games.json'))
@@ -120,15 +74,6 @@ module Store
       loaded_game.label = @labels[0]
       loaded_game.genre = @genres[0]
       @games.push(loaded_game)
-    end
-  end
-
-  def load_authors
-    puts 'Loading authors'
-    authors = JSON.parse(File.read('./data/authors.json'))['authors']
-    authors.each do |author|
-      author = Author.new author['first_name'], author['last_name']
-      @authors << author
     end
   end
